+++ src/BUPSystem/ProductGUI/Product group register.xaml.cs
@@ -14,42 +14,90 @@ namespace BUPSystem.ProductGUI
 
         public ObservableCollection<ProductGroup> GroupsList
         {
-            get
-            {
-                return new ObservableCollection<ProductGroup>(ProductGroupManagement.Instance.GetProductGroups());
-            }
+            get { return ProductGroupManagement.Instance.ProductGroups; }
+            set { ProductGroupManagement.Instance.ProductGroups = value; }
         }
-
+        /// <summary>
+        /// Standard constructor
+        /// </summary>
         public ProductGroupRegister()
         {
             InitializeComponent();
 
             DataContext = this;
         }
-
+        /// <summary>
+        /// Add a new product group
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnAdd_Click(object sender, RoutedEventArgs e)
         {
             ProductGroupCategory pgc = new ProductGroupCategory();
             pgc.ShowDialog();
+            if (pgc.DialogResult == true)
+            {
+                ProductGroupManagement.Instance.AddProductGroup(pgc.ProductGroup);
+            }
         }
-
+        /// <summary>
+        /// Edit a product group
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnChange_Click(object sender, RoutedEventArgs e)
         {
-            ProductGroupCategory pgc = new ProductGroupCategory(GroupsList[lvProductGroups.SelectedIndex]);
-            pgc.ShowDialog();
+            if (lvProductGroups.SelectedItem != null)
+            {
+                ProductGroupCategory pgc = new ProductGroupCategory(GroupsList[lvProductGroups.SelectedIndex]);
+                pgc.ShowDialog();
+                if (pgc.DialogResult == true)
+                {
+                    ProductGroupManagement.Instance.UpdateProductGroup();
+                }
+            }
+            else
+                MessageBox.Show("Markera en produktgrupp att redigera först", "Ingen vald produktgrupp");
         }
-
+        /// <summary>
+        /// Delete a product group
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnDelete_Click(object sender, RoutedEventArgs e)
         {
-            ProductGroupManagement.Instance.DeleteProductGroup(GroupsList[lvProductGroups.SelectedIndex]);
-        }
+            // Make sure the sure the user has selected an item in the listview
+            if (lvProductGroups.SelectedItem != null)
+            {
+                // Confirm that the user wishes to delete
+                MessageBoxResult mbr = MessageBox.Show("Vill du verkligen ta bort den här produktgruppen?",
+                    "Ta bort", MessageBoxButton.YesNo);
 
+                if (mbr == MessageBoxResult.Yes)
+                {
+                    ProductGroupManagement.Instance.DeleteProductGroup(GroupsList[lvProductGroups.SelectedIndex]);
+                    //lblInfo.Content = "Användaren togs bort";
+                }
+            }
+            else
+                MessageBox.Show("Markera en produktgrupp att ta bort", "Ingen vald produktgrupp");
+        }
+        /// <summary>
+        /// Select a product group for a product from productmanager
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void btnSelect_Click(object sender, RoutedEventArgs e)
         {
-            ProductGroup = GroupsList[lvProductGroups.SelectedIndex];
-            
-            DialogResult = true;
-            Close();
+            if (lvProductGroups.SelectedItem != null)
+            {
+                ProductGroup = GroupsList[lvProductGroups.SelectedIndex];
+
+                DialogResult = true;
+                Close();
+            }
+            else
+                MessageBox.Show("Markera en grupp i listan först.", "Välj en grupp");
         }
     }
 }
